fix: schemas (#33)
