+++ src/db/schema/index.ts
@@ -1,4 +1,6 @@
 export * from './shared';
 export * from './users';
+export * from './donors';
+export * from './organizations';
 export * from './needs';
 export * from './donations';
