+++ src/db/schema/shared.ts
@@ -1,23 +1,18 @@
-import { pgEnum, timestamp } from 'drizzle-orm/pg-core';
-
-export const userRole = pgEnum('user_role', ['DONOR', 'ORGANIZATION_ADMIN']);
-
-export const needStatus = pgEnum('need_status', [
-  'ACTIVE',
-  'FULFILLED',
-  'EXPIRED',
-]);
-
-export const needType = pgEnum('need_type', ['ONE_TIME']);
-
-export const donationStatus = pgEnum('donation_status', [
-  'BOOKED',
-  'FULFILLED',
-  'CANCELLED',
-]);
+import { pgTable, serial, text, timestamp, varchar } from 'drizzle-orm/pg-core';
 
 export const createdAt = () =>
   timestamp('created_at', { withTimezone: true }).notNull();
 
 export const updatedAt = () =>
   timestamp('updated_at', { withTimezone: true }).notNull();
+
+export const addresses = pgTable('addresses', {
+  id: serial('id').primaryKey(),
+  city: text('city').notNull(),
+  street: text('street').notNull(),
+  building: text('building').notNull(),
+  apartment: text('apartment'),
+  zipCode: varchar('zip_code', { length: 16 }).notNull(),
+  createdAt: createdAt(),
+  updatedAt: updatedAt(),
+});
